requirements.txt
Flask
Flask-Login
Flask-MySQLdb
